fix(component): show elapsed time within 1 month
posts older than 1 month will be show date string as before
